Adding support for posting multipart/form data and file uploads.
This commit does a number of things in order to allow file uploads to be
included in the python bindings. These are:

1. It allows subclasses of StripeObjects to specify an api_base url from
which the base endpoint will be directed. This allows a method to point
to something other than `https://api.stripe.com`.

2. By specifying `Content-Type: multipart/form-data` in the headers of a
post request, a multipart request will be created. To send a file, one
simply needs to pass a file like object as a value for one of the
parameters.

3. Creates the FileUpload resource which can be used to create and
retrieve file uploads from the Stripe upload API.