Re-use `pycurl.Curl` objects between requests
This is poorly documented, but it seems that Pycurl will maintain
established connections between requests [1]. Here we move the
initialization of the `Curl` object to the constructor and keep almsot
everything else the same, with the exception of the addition of a
`self._curl.reset()` call before any request to make sure that our
options are zeroed (`reset` doesn't reset open connections).

Fixes #324.

[1] http://pycurl.io/docs/latest/curlobject.html#pycurl.Curl.reset