Fix quotes in `version.py`
The project recently switched over to Black for code formatting in #521.
Black expects double-quoted strings throughout the project. We
double-quoted this file when that patch was merged, but our release
script was still incorrectly producing single quoted strings.

I'm just about to patch the release script, but putting this fix in here
so that we can fix `master`.