Inline StripeObject._request
Inlined StripeObject._request

got rid of e.g. "api_version = api_version or self.api_version"

No need for self.api_version as that is used only for "forwarding" the
api version when you call a method from an response object. With
services we won't do that (except for auto-pagination, which should be
handled as a special case)